Connect sign in, sign up and sign out with BE and local storage.
