+++ README.md
@@ -26,3 +26,14 @@ Formatter and lint check are run in pre commit hook. You can run it manually wit
 ### Project is deployed to mestle.cz with netlify
 
 [![Netlify Status](https://api.netlify.com/api/v1/badges/9e3e535e-ae8e-482e-afb4-9b416e0881aa/deploy-status)](https://app.netlify.com/sites/lustrous-puffpuff-c8e9da/deploys)
+
+### Issues
+
+Netlify can crash and keep port open in background during local development of netlify functions. It prevents to start server again.
+
+Windows command:
+
+```
+netstat -ano | findstr :3000.*LISTENING
+tskill port
+```
