+++ src/component/AuthForm.css
@@ -1,0 +1,63 @@
+fieldset[disabled] .button {
+  background-color: silver;
+  border-color: #a0a0a0;
+}
+
+.login-sign-up-form {
+  width: 100%;
+}
+
+.tab-group {
+  list-style: none;
+  padding: 0;
+  margin: 0 0 40px 0;
+  user-select: none;
+}
+
+.tab-group:after {
+  content: '';
+  display: table;
+  clear: both;
+}
+
+.tab-group li div {
+  display: block;
+  text-decoration: none;
+  padding: 15px;
+  background: silver;
+  color: #fff;
+  font-size: 20px;
+  float: left;
+  width: 50%;
+  text-align: center;
+  cursor: pointer;
+  transition: 0.5s ease;
+}
+
+.tab-group li div:hover {
+  background: lightskyblue;
+  color: white;
+}
+
+.tab-group li div.active {
+  background: blue;
+  color: white;
+}
+
+.field-wrap {
+  padding-bottom: 20px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  user-select: none;
+}
+
+.field-wrap input {
+  width: 80%;
+}
+
+@media screen and (min-width: 500px) {
+  .login-sign-up-form {
+    width: 50%;
+  }
+}
