+++ src/component/User.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { getScore } from '../History/history';
 import BetaSwitch from './BetaSwitch';
@@ -6,80 +6,11 @@ import { TwitterFollowButton } from 'react-twitter-embed';
 
 import './User.css';
 import { t } from '../Util/translate';
+import AuthForm from './AuthForm';
 
-const User = ({ history }) => {
-  let [authMode, setAuthMode] = useState('signin');
-
-  const changeAuthMode = () => {
-    setAuthMode(authMode === 'signin' ? 'signup' : 'signin');
-  };
-
+const User = ({ history, loggedUser, setLoggedUser }) => {
   const getUserScore = () => {
     return getScore(history);
-  };
-
-  const isSignUp = () => {
-    return authMode === 'signup';
-  };
-
-  const isSignIn = () => {
-    return authMode === 'signin';
-  };
-
-  // https://linguinecode.com/post/how-to-get-form-data-on-submit-in-reactjs
-  const [formData, updateFormData] = useState({});
-
-  const handleChange = (e) => {
-    updateFormData({
-      ...formData,
-      [e.target.name]: e.target.value.trim()
-    });
-    console.log(formData);
-  };
-
-  const signIn = async () => {
-    console.log(formData);
-    const response = await fetch('/.netlify/functions/sign_in', {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        email: formData.email,
-        password: formData.password //TODO use bcrypt
-      })
-    });
-    if (response.status === 200) {
-      setAuthMode('logged');
-    } else {
-      const data = response.json();
-      console.log('Error: ' + data);
-    }
-  };
-
-  const signUp = async () => {
-    console.log(formData);
-    const response = await fetch('/.netlify/functions/sign_up', {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username: formData.username,
-        email: formData.email,
-        password: formData.password //TODO use bcrypt
-      })
-    });
-    if (response.status === 200) {
-      setAuthMode('signin');
-    } else {
-      const data = response.json();
-      console.log('Error: ' + data);
-    }
   };
 
   return (
@@ -102,64 +33,7 @@ const User = ({ history }) => {
         {t('components.user.betaTesting.title')}
         <BetaSwitch />
       </div>
-      <div className='auth'>
-        <div className='login-sign-up-form'>
-          <ul className='tab-group'>
-            <li className='tab active'>
-              <div onClick={changeAuthMode} className={isSignIn() ? 'active' : ''}>
-                {t('components.user.login.title')}
-              </div>
-            </li>
-            <li className='tab'>
-              <div onClick={changeAuthMode} className={isSignUp() ? 'active' : ''}>
-                {t('components.user.register.title')}
-              </div>
-            </li>
-          </ul>
-          {isSignIn() && (
-            <div className='auth-form'>
-              <form>
-                <fieldset>
-                  <div className='field-wrap'>
-                    <label>{t('components.user.login.email')}</label>
-                    <input type='text' name='email' value={formData.email} autoComplete='off' onChange={handleChange} />
-                  </div>
-                  <div className='field-wrap'>
-                    <label>{t('components.user.login.password')}</label>
-                    <input type='password' name='password' value={formData.password} autoComplete='off' onChange={handleChange} />
-                  </div>
-                  <div className='button' onClick={signIn}>
-                    {t('components.user.login.submit')}
-                  </div>
-                </fieldset>
-              </form>
-            </div>
-          )}
-          {isSignUp() && (
-            <div className='auth-form'>
-              <form>
-                <fieldset>
-                  <div className='field-wrap'>
-                    <label>{t('components.user.register.username')}</label>
-                    <input type='text' name='username' value={formData.username} autoComplete='off' onChange={handleChange} />
-                  </div>
-                  <div className='field-wrap'>
-                    <label>{t('components.user.register.email')}</label>
-                    <input type='text' name='email' autoComplete='off' onChange={handleChange} />
-                  </div>
-                  <div className='field-wrap'>
-                    <label>{t('components.user.register.password')}</label>
-                    <input type='password' name='password' value={formData.password} autoComplete='off' onChange={handleChange} />
-                  </div>
-                  <div className='button' onClick={signUp}>
-                    {t('components.user.register.submit')}
-                  </div>
-                </fieldset>
-              </form>
-            </div>
-          )}
-        </div>
-      </div>
+      <AuthForm loggedUser={loggedUser} setLoggedUser={setLoggedUser} />
       <div>
         <div className='twitter-wrapper'>
           <TwitterFollowButton
@@ -177,7 +51,9 @@ const User = ({ history }) => {
 };
 
 User.propTypes = {
-  history: PropTypes.object.isRequired
+  history: PropTypes.object.isRequired,
+  loggedUser: PropTypes.object,
+  setLoggedUser: PropTypes.func.isRequired
 };
 
 export default User;
